+++ src/kannushi/_rendering/__init__.py
@@ -1,4 +1,5 @@
 __all__ = [
+    "DEFAULT_TEMPLATE_EXTENSION",
     "RenderTemplateContext",
     "RenderHandler",
     "RenderResultObserver",
@@ -15,6 +16,7 @@ __all__ = [
 #
 
 from .core import (
+    DEFAULT_TEMPLATE_EXTENSION,
     RenderTemplateContext, RenderHandler, RenderResultObserver, RenderConfig, RenderResult,
     writing_render_handler, composite_render_pipeline, validate_render_paths, render
 )